+++ src/lingua/extractors/xml.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 import collections
 import re
+import sys
 from StringIO import StringIO
 from xml.parsers import expat
 
@@ -66,8 +67,15 @@ class XmlExtractor(object):
 
         try:
             self.parser.ParseFile(fileobj)
-        except expat.ExpatError:
-            pass
+        except expat.ExpatError as e:
+            if getattr(fileobj, 'name', None):
+                print >> sys.stderr, \
+                        ('Aborting due to parse error in %s: %s' %
+                                (fileobj.name, e.message))
+            else:
+                print >> sys.stderr, \
+                        ('Aborting due to parse error: %s' % e.message)
+            sys.exit(1)
         return self.messages
 
     def addMessage(self, message, comments=[]):